requirements.txt.txt
Plaintext
streamlit
pandas
python-docx